Multiple fixes and improvements

+++ crc32.h
@@ -1,6 +1,6 @@
 #ifndef CRC32_H
 #define CRC32_H
 
-unsigned int crc32(const void *data, size_t len, unsigned int prev_crc);
+unsigned int crc32(const void *data, unsigned long long len, unsigned int prev_crc);
 
 #endif /* CRC32_H */
